requirements.txt
numpy==1.24.2
pandas==2.0.0
seaborn==0.12.2
matplotlib==3.7.1
scikit-learn==1.2.2
requests==2.26.0
pyyaml==5.4.1
boto3==1.20.0
pathlib==1.0.1
pathlib2==2.3.7.post1


